Git prompt edited

+++ install_vim_ide.sh
@@ -1,21 +1,32 @@
 #!/bin/bash
 
-set -e 
+#set -e 
+deleted=`pwd`/d_ctags
 
-
+function run_once
+{
+    if [ -f !d_ctags ];then
+        rm -rf /usr/bin/ctags
+        rm -rf /usr/bin/etags
+        touch d_ctags
+        sleep 20
+    fi
+}
 function vim_ide_packages
 {
     apt-get install -y vim-gnome ruby-dev bash-completion 
-    rm -rf /usr/bin/ctags
-    rm -rf /usr/bin/etags
-    sleep 20
+    run_once
+    
     pip install dbgp vim-debug
     install-vim-debug.py
     cp -r ./* $HOME/.vim
     cp -r ./.vimrc $HOME
     apt-get install -y exuberant-ctags
-    sed -i "$ aexport GIT_PS1_SHOWDIRTYSTATE=1" >> $HOME/.bashrc
-    sed -i "$ aexport PS1='\[\033[0;34m\]\u@\w\[\033[0;32m\]$(__git_ps1)\[\033[0;34m\]\$\[\033[0m\] '" >> $HOME/.bashrc
+    sleep 5
+    echo "setting bashrc"
+    cat git_prompt >> $HOME/.bashrc
+    sleep 5
+    source ~/.bashrc
    
 }
